+++ tests/testthat/test-agvgd.R
@@ -36,7 +36,7 @@ test_that("align01", {
 
   example <- '01'
   agvgdweb_output_file <- paste0('alignments/agvgd', example, '.tsv')
-  agvgdweb <- import_agvgdweb(agvgdweb_output_file)
+  agvgdweb <- read_agvgdweb_results(agvgdweb_output_file)
 
   alignment_file <- paste0('alignments/align', example, '.fa')
   alignment <- read_alignment(file = alignment_file)
@@ -45,14 +45,14 @@ test_that("align01", {
 
   # agvgd[-2]: drops col 2, i.e., the `poi` which is non-existent in the
   # agvgdweb output
-  expect_equal(agvgdweb, agvgd[-2], tolerance = 0.001)
+  expect_equal(agvgdweb[-2], agvgd[-2], tolerance = 0.001)
 })
 
 test_that("align02", {
 
   example <- '02'
   agvgdweb_output_file <- paste0('alignments/agvgd', example, '.tsv')
-  agvgdweb <- import_agvgdweb(agvgdweb_output_file)
+  agvgdweb <- read_agvgdweb_results(agvgdweb_output_file)
 
   alignment_file <- paste0('alignments/align', example, '.fa')
   alignment <- read_alignment(file = alignment_file)
@@ -61,14 +61,14 @@ test_that("align02", {
 
   # agvgd[-2]: drops col 2, i.e., the `poi` which is non-existent in the
   # agvgdweb output
-  expect_equal(agvgdweb, agvgd[-2], tolerance = 0.001)
+  expect_equal(agvgdweb[-2], agvgd[-2], tolerance = 0.001)
 })
 
 test_that("align03", {
 
   example <- '03'
   agvgdweb_output_file <- paste0('alignments/agvgd', example, '.tsv')
-  agvgdweb <- import_agvgdweb(agvgdweb_output_file)
+  agvgdweb <- read_agvgdweb_results(agvgdweb_output_file)
 
   alignment_file <- paste0('alignments/align', example, '.fa')
   alignment <- read_alignment(file = alignment_file)
@@ -77,14 +77,14 @@ test_that("align03", {
 
   # agvgd[-2]: drops col 2, i.e., the `poi` which is non-existent in the
   # agvgdweb output
-  expect_equal(agvgdweb, agvgd[-2], tolerance = 0.001)
+  expect_equal(agvgdweb[-2], agvgd[-2], tolerance = 0.001)
 })
 
 test_that("align04", {
 
   example <- '04'
   agvgdweb_output_file <- paste0('alignments/agvgd', example, '.tsv')
-  agvgdweb <- import_agvgdweb(agvgdweb_output_file)
+  agvgdweb <- read_agvgdweb_results(agvgdweb_output_file)
 
   alignment_file <- paste0('alignments/align', example, '.fa')
   alignment <- read_alignment(file = alignment_file)
@@ -93,14 +93,14 @@ test_that("align04", {
 
   # agvgd[-2]: drops col 2, i.e., the `poi` which is non-existent in the
   # agvgdweb output
-  expect_equal(agvgdweb, agvgd[-2], tolerance = 0.001)
+  expect_equal(agvgdweb[-2], agvgd[-2], tolerance = 0.001)
 })
 
 test_that("align05", {
 
   example <- '05'
   agvgdweb_output_file <- paste0('alignments/agvgd', example, '.tsv')
-  agvgdweb <- import_agvgdweb(agvgdweb_output_file)
+  agvgdweb <- read_agvgdweb_results(agvgdweb_output_file)
 
   alignment_file <- paste0('alignments/align', example, '.fa')
   alignment <- read_alignment(file = alignment_file)
@@ -109,14 +109,14 @@ test_that("align05", {
 
   # agvgd[-2]: drops col 2, i.e., the `poi` which is non-existent in the
   # agvgdweb output
-  expect_equal(agvgdweb, agvgd[-2], tolerance = 0.001)
+  expect_equal(agvgdweb[-2], agvgd[-2], tolerance = 0.001)
 })
 
 test_that("align06", {
 
   example <- '06'
   agvgdweb_output_file <- paste0('alignments/agvgd', example, '.tsv')
-  agvgdweb <- import_agvgdweb(agvgdweb_output_file)
+  agvgdweb <- read_agvgdweb_results(agvgdweb_output_file)
 
   alignment_file <- paste0('alignments/align', example, '.fa')
   alignment <- read_alignment(file = alignment_file)
@@ -125,14 +125,14 @@ test_that("align06", {
 
   # agvgd[-2]: drops col 2, i.e., the `poi` which is non-existent in the
   # agvgdweb output
-  expect_equal(agvgdweb, agvgd[-2], tolerance = 0.001)
+  expect_equal(agvgdweb[-2], agvgd[-2], tolerance = 0.001)
 })
 
 test_that("align07", {
 
   example <- '07'
   agvgdweb_output_file <- paste0('alignments/agvgd', example, '.tsv')
-  agvgdweb <- import_agvgdweb(agvgdweb_output_file)
+  agvgdweb <- read_agvgdweb_results(agvgdweb_output_file)
 
   alignment_file <- paste0('alignments/align', example, '.fa')
   alignment <- read_alignment(file = alignment_file)
@@ -141,14 +141,14 @@ test_that("align07", {
 
   # agvgd[-2]: drops col 2, i.e., the `poi` which is non-existent in the
   # agvgdweb output
-  expect_equal(agvgdweb, agvgd[-2], tolerance = 0.001)
+  expect_equal(agvgdweb[-2], agvgd[-2], tolerance = 0.001)
 })
 
 test_that("align08", {
 
   example <- '08'
   agvgdweb_output_file <- paste0('alignments/agvgd', example, '.tsv')
-  agvgdweb <- import_agvgdweb(agvgdweb_output_file)
+  agvgdweb <- read_agvgdweb_results(agvgdweb_output_file)
 
   alignment_file <- paste0('alignments/align', example, '.fa')
   alignment <- read_alignment(file = alignment_file)
@@ -157,14 +157,14 @@ test_that("align08", {
 
   # agvgd[-2]: drops col 2, i.e., the `poi` which is non-existent in the
   # agvgdweb output
-  expect_equal(agvgdweb, agvgd[-2], tolerance = 0.001)
+  expect_equal(agvgdweb[-2], agvgd[-2], tolerance = 0.001)
 })
 
 test_that("align09", {
 
   example <- '09'
   agvgdweb_output_file <- paste0('alignments/agvgd', example, '.tsv')
-  agvgdweb <- import_agvgdweb(agvgdweb_output_file)
+  agvgdweb <- read_agvgdweb_results(agvgdweb_output_file)
 
   alignment_file <- paste0('alignments/align', example, '.fa')
   alignment <- read_alignment(file = alignment_file)
@@ -173,6 +173,6 @@ test_that("align09", {
 
   # agvgd[-2]: drops col 2, i.e., the `poi` which is non-existent in the
   # agvgdweb output
-  expect_equal(agvgdweb, agvgd[-2], tolerance = 0.001)
+  expect_equal(agvgdweb[-2], agvgd[-2], tolerance = 0.001)
 })